# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Sumi Agentic Voice CRM API
  version: 1.0.0
  description: Contract-first text/audio CRM interaction API. Provider implementations are replaceable adapters.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /v1/ask:
    post:
      operationId: ask
      parameters:
        - { $ref: '#/components/parameters/TenantId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AskRequest' }
          multipart/form-data:
            schema:
              type: object
              required: [audio, metadata]
              properties:
                audio: { type: string, format: binary }
                metadata: { type: string }
      responses:
        '200': { description: Completed, content: { application/json: { schema: { $ref: '#/components/schemas/AskResponse' } } } }
        '202': { description: Review required, content: { application/json: { schema: { $ref: '#/components/schemas/ReviewResponse' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '415': { $ref: '#/components/responses/Error' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
        '504': { $ref: '#/components/responses/Error' }
  /v1/tts/synthesize:
    post:
      operationId: synthesize
      security:
        - bearerAuth: []
      parameters:
        - { $ref: '#/components/parameters/TenantId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TtsRequest' } } }
      responses:
        '201': { description: Audio asset, content: { application/json: { schema: { $ref: '#/components/schemas/TtsAsset' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    TenantId: { name: X-Tenant-Id, in: header, required: true, schema: { type: string, minLength: 1 } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 8, maxLength: 128 } }
  responses:
    Error: { description: Error envelope, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
    TextAskRequest:
      type: object
      required: [input, output_mode]
      properties:
        input: { type: object, required: [type, text], properties: { type: { const: text }, text: { type: string, minLength: 1, maxLength: 10000 } } }
        output_mode: { type: string, enum: [text, audio, both] }
        locale: { type: string, enum: [zh-CN, en-US, hi-IN, te-IN], default: zh-CN }
        conversation_id: { type: string }
    AudioAskRequest:
      type: object
      required: [input, output_mode]
      properties:
        input:
          type: object
          required: [type, data_base64, content_type]
          properties:
            type: { const: audio }
            data_base64: { type: string, contentEncoding: base64, minLength: 1 }
            content_type: { type: string, enum: [audio/wav, audio/x-wav, audio/mpeg, audio/mp4, audio/webm] }
        output_mode: { type: string, enum: [text, audio, both] }
        locale: { type: string, enum: [zh-CN, en-US, hi-IN, te-IN], default: zh-CN }
    AskRequest:
      oneOf:
        - { $ref: '#/components/schemas/TextAskRequest' }
        - { $ref: '#/components/schemas/AudioAskRequest' }
      discriminator: { propertyName: input.type }
    Understanding:
      type: object
      required: [intent, confidence, entities, missing, needs_confirmation, schema_version]
      properties:
        intent: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
        entities: { type: object, additionalProperties: true }
        missing: { type: array, items: { type: string } }
        needs_confirmation: { type: boolean }
        schema_version: { type: string }
    AskResponse:
      type: object
      required: [request_id, status, input, understanding, answer]
      properties:
        request_id: { type: string }
        status: { const: completed }
        input: { type: object, additionalProperties: true }
        understanding: { $ref: '#/components/schemas/Understanding' }
        crm: { type: object, additionalProperties: true }
        answer: { type: object, required: [text], properties: { text: { type: string }, language: { type: string } } }
        audio: { $ref: '#/components/schemas/TtsAsset' }
    ReviewResponse:
      type: object
      required: [request_id, status, understanding, review_task]
      properties:
        request_id: { type: string }
        status: { const: needs_review }
        understanding: { $ref: '#/components/schemas/Understanding' }
        review_task: { type: object, required: [id], properties: { id: { type: string }, expires_at: { type: string, format: date-time } } }
    TtsRequest:
      type: object
      required: [text, language, format]
      properties:
        text: { type: string, minLength: 1, maxLength: 5000 }
        language: { type: string, enum: [zh-CN, en-US, hi-IN, te-IN] }
        voice: { type: string }
        format: { type: string, enum: [mp3, wav, ogg] }
    TtsAsset:
      type: object
      required: [asset_id, url, mime_type, status]
      properties:
        asset_id: { type: string }
        url: { type: string, format: uri-reference }
        mime_type: { type: string }
        duration_ms: { type: integer, minimum: 0 }
        expires_at: { type: string, format: date-time }
        status: { type: string, enum: [ready, failed, pending] }
    ErrorEnvelope:
      type: object
      required: [request_id, status, error]
      properties:
        request_id: { type: string }
        status: { const: failed }
        error: { type: object, required: [code, message, retryable], properties: { code: { type: string }, message: { type: string }, retryable: { type: boolean }, details: { type: object } } }